src/gateway/requirements.txt
click==8.1.3
grpcio==1.51.1
grpcio-tools==1.51.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
protobuf==4.21.12
Werkzeug==2.2.2
